requirements.txt
langchain
langchain-anthropic
langchain-community
langchain-docling
langchain_google_genai
langchain-huggingface
langchain-ollama
langchain-openai
langchain-unstructured
openai
python-dotenv
ipywidgets
IPython
ipykernel
streamlit
pymysql
pypdf
pymupdf
pdfplumber
bs4
arxiv
fitz
python-magic-bin
unstructured
unstructured[pdf]